Changing count check
Like the other test I fixed earlier, 2 queries are being issued to do the initial fetch (possibly a sub limit query or something, I made a note to investigate a bit to be sure it's not something sinister).  I believe these count checks were put in place to be sure that relationship fetching wasn't causing additional queries if the relationship is defined in the select.  By moving the count, I believe this check is still valid.