Adding explicit order clause
Not sure why the order returned for me is different than what it was when these tests were written (possibly the way that newer versions of sqlite return rows if no order is specified), but added explicit order clauses to a couple of test cases, then re-ordered the expectations to match. Tests pass now.