Doctrine_Record_Filter_Compound PHP 7 fix
`isset($record[$alias][$name]))` behaves differently in PHP 5 and 7:

* In PHP 7 it results in two `\Doctrine_Access::contains()` calls, one for the `$alias` and then one for `$name`.
* In PHP 5, there's a call to `\Doctrine_Access::get()` for $alias and then a `\Doctrine_Access::contains()` for the `$name`.

However, there is no `\Doctrine_Record_Filter::filterContains()` method, which means that we don't have a good way of returning true for `$alias` call to `contains()`. It seems acceptable to just change it get the relation first, as this is similar to the PHP 5 isset behavior anyway.